win: change ssize_t typedef to SSIZE_T
libmaxminddb redefines this when building under MSVC and breaks the
build otherwise due to the redefinition of ssize_t

+++ win/inspircd_win32wrapper.h
@@ -85,7 +85,7 @@
 #define strcasecmp _stricmp
 #define strncasecmp _strnicmp
 
-typedef int ssize_t;
+typedef SSIZE_T ssize_t;
 
 /* _popen, _pclose */
 #define popen _popen
